IFEM.V8.2.2008
Xingshi Wang, 06/03/2008
Add options for FEM interpolation
- add additional option in input_solid.in for FEM interpolation, RKPM=1,
  FEM=2;
- added search_3d.f90,search_inf.f90,sh_exchange.f90,migs.f90
-----------------------------------------------------------------
Xingshi Wang, 06/03/2008
Add shx_tets.f90 to calculate 3-D tets case's shapefunction in
data_exchange.f90.
-------------------------------------------------
Xingshi Wang, 05/31/2008
1) In r_sstif.f90 change 'acc=acc-xmg', so that the gravity term is
consistent with the eqaution.

2) The internal force should be in the opposite direction when it is
distributed onto the fluid nodes. Therefore, I just change the sign in
r_sstif.f90 instead of change the sign in the interpolation codes since the
external and acceleration terms should keep their signs. In r_sstif.f90, it is
changed to FSI=-interal+exteranl-acceleration.

-----------------------------------------------------
Xingshi Wang, 05/09/2008
In r_spiola_elastic to calculate first Piola-Kirchhoff stress, "J" should be
the determinant of the deformation gradient matrix, but originally we use det
which is the determinant of Jacobi to initial coordinates.
define "detf" and call determinant to solve the determinant of xot which is
the inverse of F.  detf=1/{det(F)}

Advised-by: Lucy T. Zhang <zhangLucy@rpi.edu>